fix: cast quality_profile_id to i32 for Postgres bind
The search query param is i64 but quality_profiles.id is SERIAL (i32).
sqlx sends INT8 wire format which Postgres rejects with "incorrect
binary data format in bind parameter 1".

+++ crates/stackarr-web/src/routes/releases.rs
@@ -83,7 +83,7 @@ async fn search_releases(
             match sqlx::query_as::<_, QualityProfile>(
                 "SELECT * FROM quality_profiles WHERE id = $1",
             )
-            .bind(id)
+            .bind(id as i32)
             .fetch_optional(pool)
             .await
             {
